// src/app/api/civitai.ts
export enum ModelTypes {
    "Checkpoint",
    "TextualInversion",
    "Hypernetwork",
    "AestheticGradient",
    "LORA",
    "Controlnet",
    "Poses"
}
// ...
export type ModelStats = {
    downloadCount: number;
    favoriteCount: number;
    commentCount: number;
    ratingCount: number;
    rating: number;
}

export type ModelHashes = {
    AutoV2: string;
    SHA256: string;
    CRC32: string;
    BLAKE3: string;
}

export type ModelVersionFile = {
    name:string;
    id: number;
    sizeKB: number;
    type: string;
    metadata: {
        fp: string;
        size: string;
        format: string;
    };
    pickleScanResult: string;
    pickleScanMessage: string;
    virusScanResult: string;
    scannedAt: string;
    hashes: ModelHashes;
    downloadUrl: string;
    primary: boolean;
}

export type ModelImageMeta = {
    ENSD: string;
    Size: string;
    seed: number;
    Score: string;
    steps: number;
    prompt: string;
    sampler: string;
    "Eta DDIM": string;
    cfgScale: number;
    resources: [];
    "Model hash": string;
    "Hires upscale": string;
    "Hires upscaler": string;
    negativePrompt: string;
    "Denoising strength": string;
}

export type ModelImagesItem = {
    url: string;
    nsfw: boolean;
    width: number;
    height: number;
    hash: string;
    meta: ModelImageMeta;
}

export type ModelVersionsItem = {
    id: number;
    modelId: number;
    name: string;
    createdAt: string;
    updatedAt: string;
    trainedWords: string[];
    baseModel: string;
    earlyAccessTimeFrame: number;
    description: string;
    stats: {
        downloadCount: number;
        ratingCount: number;
        rating: number;
    }
    files: ModelVersionFile[];
    images: ModelImagesItem[];
    downloadUrl: string;
}

export type CivitAIModelItem = {
    id: number;
    name: string;
    type: ModelTypes;
    poi: boolean;
    nsfw: boolean;
    allowNoCredit: boolean;
    allowCommercialUse: boolean;
    allowDerivatives: boolean;
    allowDifferentLicense: boolean;
    stats: ModelStats;
    creator: {
        username: string;
        image: string;
    }
    tags: string[];
    modelVersions: ModelVersionsItem[];
}